ml/requirements.txt
librosa==0.8.1
scipy==1.4.1
numpy==1.18.5
matplotlib==3.0.2
Pillow==5.4.1
torch
opencv-python